Added RealTrashBin.
Updated some return values in class diagram

+++ Code/TCleaner/protocol/handlers/BatteryBoardPacketHandler.h
@@ -42,7 +42,7 @@ class BatteryBoardPacketHandler : public protocol::BoardPacketHandler
 		char boardid;
 		PacketServer * ps;
 		
-		double currentValue;
+		int currentValue;
 		bool full;
 		bool empty;
 };
